adds modules
